export to `StampComparisonResult` to util to avoid circular import
